requirements.txt
flask==2.3.3
flask-jwt-extended==4.5.2
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.0.74
numpy==1.24.3
gunicorn==21.2.0
pytest==7.4.2
python-jose==3.3.0
requests==2.31.0
pillow==10.0.0
flask-cors==4.0.0
bcrypt==4.0.1
face-recognition==1.3.0
dlib==19.24.0
